finish authorization, fix updating

+++ src/restshop/api/resources/NarociloResource.java
@@ -10,22 +10,32 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriBuilder;
 
 import restshop.dao.NarociloDAO;
+import restshop.dao.StanjeDAO;
+import restshop.dao.UporabnikDAO;
 import restshop.entities.Narocilo;
+import restshop.entities.Uporabnik;
 import restshop.entities.lists.NarociloList;
 
 @Path("/narocila")
 public class NarociloResource extends Resource<Narocilo> {
 	
 	NarociloDAO ndao=new NarociloDAO();
+	UporabnikDAO udao=new UporabnikDAO();
+	StanjeDAO sdao=new StanjeDAO();
 
 	@POST
 	@Consumes({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
-	public Response create(Narocilo entity) {
+	public Response create(@Context SecurityContext sc, Narocilo entity) {
+		Uporabnik u=udao.read(sc.getUserPrincipal().getName());
+		entity.setUporabnik(u);
+		entity.setStanje(sdao.read("kosarica"));
 		ndao.create(entity);
 		UriBuilder ub=uriInfo.getBaseUriBuilder();
 		URI uri=ub.path(NarociloResource.class).path(Integer.toString(entity.getId_narocilo())).build();
@@ -35,10 +45,20 @@ public class NarociloResource extends Resource<Narocilo> {
 	@GET
 	@Path("/{id}")
 	@Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
-	public Response read(@PathParam("id") int id) {
+	public Response read(@Context SecurityContext sc, @PathParam("id") int id) {
 		Narocilo entity=ndao.read(id);
 		if(entity!=null) {
-			return Response.ok(entity).build();
+			if(sc.isUserInRole("admin")) {
+				return Response.ok(entity).build();
+			} else if(sc.isUserInRole("uporabnik")) {
+				if(entity.getUporabnik().getUp_ime().equals(sc.getUserPrincipal().getName())) {
+					return Response.ok(entity).build();
+				} else {
+					return Response.status(404).build();
+				}
+			} else {
+				return Response.status(404).build();
+			}
 		} else {
 			return Response.status(404).build();
 		}
@@ -47,9 +67,21 @@ public class NarociloResource extends Resource<Narocilo> {
 	@PUT
 	@Path("/{id}")
 	@Consumes({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
-	public Response update(@PathParam("id") int id, Narocilo entity) {
+	public Response update(@Context SecurityContext sc, @PathParam("id") int id, Narocilo entity) {
 		entity.setId_narocilo(id);
-		boolean updated=ndao.update(entity);
+		Narocilo old=ndao.read(id);
+		boolean updated=false;
+		if(sc.isUserInRole("admin")) {
+			updated=ndao.update(entity);
+		} else if(sc.isUserInRole("uporabnik")) {
+			if(sc.getUserPrincipal().getName().equals(old.getUporabnik().getUp_ime())) {
+				if(old.getStanje().getId_stanje()==1 && entity.getStanje().getId_stanje()==2) {
+					updated=ndao.update(entity);
+				} else {
+					return Response.ok().entity("State invalid or missing").build();
+				}
+			}
+		}
 		if(updated) {
 			return Response.ok().entity("Resource updated").build();
 		} else {
@@ -59,8 +91,20 @@ public class NarociloResource extends Resource<Narocilo> {
 
 	@DELETE
 	@Path("/{id}")
-	public Response delete(@PathParam("id") int id) {
-		boolean deleted=ndao.delete(id);
+	public Response delete(@Context SecurityContext sc, @PathParam("id") int id) {
+		boolean deleted=false;
+		if(sc.isUserInRole("admin")) {
+			deleted=ndao.delete(id);
+		} else if(sc.isUserInRole("uporabnik")) {
+			Narocilo n=ndao.read(id);
+			if(sc.getUserPrincipal().getName().equals(n.getUporabnik().getUp_ime())) {
+				if(n.getStanje().getId_stanje()==1) {
+					deleted=ndao.delete(id);
+				} else {
+					return Response.status(400).entity("Cannot delete this resource").build();
+				}
+			}
+		}
 		if(deleted) {
 			return Response.ok().entity("Resource deleted").build();
 		} else {
@@ -70,9 +114,18 @@ public class NarociloResource extends Resource<Narocilo> {
 	
 	@GET
 	@Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
-	public Response list() {
-		NarociloList list=new NarociloList(ndao.list());
-		return Response.ok(list).build();
+	public Response list(@Context SecurityContext sc) {
+		NarociloList list=null;
+		if(sc.isUserInRole("admin")) {
+			list=new NarociloList(ndao.list());
+			return Response.ok(list).build();
+		} else if(sc.isUserInRole("uporabnik")) {
+			Uporabnik u=udao.read(sc.getUserPrincipal().getName());
+			list=new NarociloList(ndao.list(u.getId_uporabnik()));
+			return Response.ok(list).build();
+		} else {
+			return null;
+		}
 	}
 
 }
